Add support for Windows (#9)
Build error in Windows is caused by &PyType_Type which is a PyObject* as the type of a type object is
'type', but this isn’t strictly conforming C and some compilers, e.g Visual C and MinGW's gcc, complain.

So we change:

PyVarObject_HEAD_INIT(&PyType_Type, 0)

to:

PyVarObject_HEAD_INIT(NULL, 0)

Fortunately, this member will be filled in for us by PyType_Ready() making everything work.